refacto: change some archi of tests
